Dockers/Docker Compose/flask_app/requirements.txt
Flask==2.0.3
psycopg2-binary==2.9.3
Werkzeug==2.0.3  # Ensure compatibility with Flask